// frontend/src/types/gtfs.d.ts
import type { RouteType } from "./gtfsCustom";

// Autogenerated by generate-types.ts
// Modified to add RouteType

export type Agency = {
    agencyId: (string | null),
    agencyName: string,
    agencyUrl: string,
    agencyTimezone: string,
    agencyLang: (string | null),
    agencyPhone: (string | null),
    agencyFareUrl: (string | null),
    agencyEmail: (string | null)
};

export type CalendarDate = {
    serviceId: string,
    date: Date,
    exceptionType: number,
    holidayName: (string | null)
};

export type FeedInfo = {
    feedPublisherName: string,
    feedPublisherUrl: string,
    feedLang: string,
    defaultLang: (string | null),
    feedStartDate: (Date | null),
    feedEndDate: (Date | null),
    feedVersion: (string | null),
    feedContactEmail: (string | null),
    feedContactUrl: (string | null)
};

export type Pathway = {
    pathwayId: string,
    fromStopId: string,
    toStopId: string,
    pathwayMode: number,
    isBidirectional: boolean,
    length: (number | null),
    traversalTime: (number | null),
    stairCount: (number | null),
    maxSlope: (number | null),
    minWidth: (number | null),
    signpostedAs: (string | null),
    reversedSignpostedAs: (string | null)
};

export type Route = {
    routeId: string,
    agencyId: (string | null),
    routeShortName: (string | null),
    routeLongName: (string | null),
    routeDesc: (string | null),
    routeType: RouteType,
    routeUrl: (string | null),
    routeColor: (string | null),
    routeTextColor: (string | null),
    routeSortOrder: (number | null),
    continuousPickup: (number | null),
    continuousDropoff: (number | null),
    networkId: (string | null)
};

export type Shape = {
    shapeId: string,
    shapePtLat: number,
    shapePtLon: number,
    shapePtSequence: number,
    shapeDistTravaled: (number | null)
};

export type StopTime = {
    tripId: (string | null),
    arrivalTime: (string | null),
    arrivalTimestamp: (Date | null),
    departureTime: (string | null),
    departureTimestamp: (Date | null),
    locationGroupId: (string | null),
    locationId: (string | null),
    stopId: string,
    stopSequence: number,
    stopHeadsign: (string | null),
    startPickUpDropOffWindow: (string | null),
    startPickUpDropOffWindowTimestamp: (Date | null),
    endPickUpDropOffWindow: (string | null),
    endPickUpDropOffWindowTimestamp: (Date | null),
    pickupType: (number | null),
    dropOffType: (number | null),
    continuousPickup: (number | null),
    continuousDropOff: (number | null),
    shapeDistTraveled: (number | null),
    timepoint: (number | null),
    pickupBookingRuleId: (string | null),
    dropOffBookingRuleId: (string | null)
};

export type Stop = {
    stopId: string,
    stopCode: (string | null),
    stopName: (string | null),
    ttsStopName: (string | null),
    stopDesc: (string | null),
    stopLat: (number | null),
    stopLon: (number | null),
    zoneId: (string | null),
    stopUrl: (string | null),
    locationType: (number | null),
    parentStation: (string | null),
    stopTimezone: (string | null),
    wheelchairBoarding: (number | null),
    levelId: (string | null),
    platformCode: (string | null)
};

export type Trip = {
    routeId: string,
    serviceId: string,
    tripId: string,
    tripHeadsign: (string | null),
    tripShortName: (string | null),
    directionId: (number | null),
    blockId: (string | null),
    shapeId: (string | null),
    wheelchairAccessible: (number | null),
    bikesAllowed: (number | null)
};